requirements.txt
# TODO: Update with exact module version
paddlepaddle-gpu
numpy
opencv_python
loguru
scikit-image
tqdm
pillow
thop
ninja
tabulate
tensorboard

